IDs completo

+++ src/Main.scala
@@ -95,25 +95,25 @@ object Main {
 
   def loadSong(line: String): Unit={
     val info=line.split(";").toList
-    val song:Song = Song(info(0).toInt,info(1),info(2),info(3).toInt,info(4).toInt,info(5),info(6).toInt,info(7),info(8).toInt,info(9).toInt)
+    val song:Song = Song(info(0),info(1),info(2),info(3),info(4),info(5),info(6),info(7),info(8),info(9))
     loadedSongs+= song
     println("Song loaded from DB")
   }
   def loadArtist(line: String): Unit={
     val info=line.split(";").toList
-    val artist:Artist = Artist(info(0).toInt,info(1),info(2),info(3))
+    val artist:Artist = Artist(info(0),info(1),info(2),info(3))
     loadedArtists+= artist
     println("Artist loaded from DB")
   }
   def loadAlbum(line: String): Unit={
     val info=line.split(";").toList
-    val album:Album = Album(info(0).toInt,info(1),info(2),info(3).toInt)
+    val album:Album = Album(info(0),info(1),info(2),info(3))
     loadedAlbums+=album
     println("Album loaded from DB")
   }
   def loadPlaylist(line: String): Unit={
     val info=line.split(";").toList
-    val playlist:Playlist =Playlist(info(0).toInt,info(1),info(2),info(3))
+    val playlist:Playlist =Playlist(info(0),info(1),info(2),info(3))
     loadedPlaylists+=playlist
     println("Playlist loaded from DB")
   }
